+++ .github/workflows/github-actions.yml
@@ -120,3 +120,4 @@ jobs:
             sudo docker pull ${{ secrets.DOCKER_REPONAME }}/${{ secrets.DOCKER_IMAGE }}
             sudo docker run -d -p 8080:8080 ${{ secrets.DOCKER_REPONAME }}/${{ secrets.DOCKER_IMAGE }}
             sudo docker image prune -f
+            sudo docker ps
