+++ src/main/java/com/sparta/kidscafe/common/config/WebConfig.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -28,5 +29,14 @@ public class WebConfig implements WebMvcConfigurer {
 	public void addInterceptors(InterceptorRegistry registry) {
 		registry.addInterceptor(jwtValidInterceptor)
 			.excludePathPatterns("/api/oauth2/**", "/api/auth/**", "/favicon.ico", "/error", "/index.html", "/",  "/redirect/oauth/**");
+	}
+
+	@Override
+	public void addCorsMappings(CorsRegistry registry) {
+		registry.addMapping("/**") // 모든 경로 허용
+				.allowedOrigins("http://localhost:3000", "http://localhost:8080") // 허용할 출처
+				.allowedMethods("GET", "POST", "PUT", "DELETE") // 허용할 HTTP 메서드
+				.allowedHeaders("*") // 모든 헤더 허용
+				.allowCredentials(true); // 인증 정보 허용
 	}
 }
